Backend/requirements.txt
# ========================================
# GURUKUL PLATFORM - COMPLETE DEPENDENCIES
# ========================================

# Core Python Dependencies
pip>=23.0.0
setuptools>=65.0.0
wheel>=0.38.0

# Data Science & Analytics
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Advanced Forecasting (Prophet & ARIMA)
prophet>=1.1.4
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Machine Learning & AI
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.20.0
sentence-transformers>=2.2.0
# Local Model Support (Arabic Model Integration)
peft>=0.17.1
bitsandbytes>=0.41.0
accelerate>=0.20.0
# Optional: unsloth (for optimized training/inference utilities)
# unsloth>=2024.1

# LangChain Framework (Financial Simulator)
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.30
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# Vector Search & Embeddings (Orchestration System)
faiss-cpu>=1.7.4
chromadb>=0.4.0
pinecone-client>=2.2.0
weaviate-client>=3.15.0

# Web Frameworks & APIs
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
requests>=2.28.0
flask>=2.2.0
flask-cors>=4.0.0
gunicorn>=20.1.0

# Database & Caching
redis>=4.5.0
celery>=5.2.0
sqlalchemy>=1.4.0

# Computer Vision (Base Backend)
opencv-python>=4.7.0
pillow>=9.4.0
easyocr>=1.7.0
pytesseract>=0.3.10

# Audio & TTS
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.1.0
soundfile>=0.12.0
librosa>=0.10.0
speechrecognition>=3.10.0

# AI API Clients
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.0

# Communication & Networking
websockets>=10.4
aiohttp>=3.8.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jsonschema>=4.17.0
psutil>=5.9.0
schedule>=1.2.0
python-multipart>=0.0.6

# Cloud Storage (for checkpoint downloads)
boto3>=1.28.0  # AWS S3 support for Arabic model checkpoint
requests>=2.31.0  # For direct URL downloads

# Development & Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional Utilities
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dateutil>=2.8.0
pytz>=2022.7
tqdm>=4.64.0

# Financial Analysis (Financial Simulator)
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Graph Processing (LangGraph)
networkx>=3.0
graphviz>=0.20.0

# Memory Management
pymongo>=4.3.0
chromadb>=0.4.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.0.0
